ai-engine/requirements.txt
fastapi==0.109.2
uvicorn[standard]==0.27.1
pydantic==2.11.4
python-dotenv==1.0.1
websockets==12.0
openai==1.79.0
openai-agents
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
sqlalchemy[asyncio]==2.0.41
pytz==2025.2
requests
PyJWT[crypto]>=2.10.1
redis==6.1.0
asyncpg==0.30.0
sentry-sdk[fastapi]==2.29.1
httpx==0.28.1
youtube-transcript-api==0.1.6